requirements.txt
pycryptodome
YamlInfo
click